undo progress - stack issues

+++ src/ecs/Components/Position.java
@@ -2,11 +2,16 @@ package ecs.Components;
 
 public class Position extends Component{
 
-    public float i;
-    public float j;
+    public int i;
+    public int j;
 
     public Position(int i, int j){
         this.i = i;
         this.j = j;
     }
+
+//    @Override
+//    public Position clone(){
+//        return new Position(this.i, this.j);
+//    }
 }
